+++ src/angular-wurfl-image-tailor.js
@@ -1,5 +1,5 @@
 /*
- * angular-wurfl-image-tailor v0.9.3
+ * angular-wurfl-image-tailor
  * Authors: Luca Corbo (https://github.com/lucor)
  * (c) 2014 - 2015 ScientiaMobile, Inc.
  * License: MIT
@@ -24,30 +24,37 @@ angular.module('angular-wurfl-image-tailor', [])
         return {
             restrict: 'E',
             replace: false,
-            scope:{},
-            template: function(element, attributes){
-                if ('ngSrc' in attributes) {
-                    return '<div class="wit"><img ng-src="{{wit_link}}"/></div>';
-                } else {
-                    return '<div class="wit"><img src="{{wit_link}}"/></div>';
-                }
+            scope: {
+                src: '@',
+                w: '@',
+                h: '@',
+                pc: '@',
+                m: '@',
+                f: '@',
+                r: '@'
+            },
+            template: function (element, attributes) {
+                return '<div class="wit"><img ng-src="{{wit_link}}"/></div>';
             },
             link: function (scope, element, attributes) {
-                var srcAName = 'ngSrc' in attributes ? 'ngSrc' : 'src';
                 scope.wit_link = '';
 
-                attributes.$observe(srcAName, function (src) {
-                    if(!src) {
-                        scope.wit_link = '';
+                var allowedAttributes = ['w', 'h', 'pc', 'm', 'f', 'r'];
+
+                scope.$watchCollection('[src, w, h, pc, m, f, r]', function (values, oldValues) {
+
+                    var src = values.shift();
+                    if (!src) {
                         return;
                     }
 
                     var wit_link_pieces = [witUrls.get()];
-                    angular.forEach(attributes['$attr'], function (attr) {
-                        if (attr != 'src' && attr != 'ng-src') {
-                            wit_link_pieces.push(attr + '_' + attributes[attr]);
+                    angular.forEach(allowedAttributes, function (attribute, index) {
+                        if (values[index]) {
+                            wit_link_pieces.push(attribute + '_' + values[index]);
                         }
                     });
+
                     wit_link_pieces.push(src);
                     scope.wit_link = wit_link_pieces.join('/');
                 });